refactor: dto 수정
#499
